+++ crispy_forms/templatetags/crispy_forms_utils.py
@@ -2,6 +2,7 @@
 import re
 
 from django import template
+from django.conf import settings
 try:  # Django < 1.4
     from django.utils.encoding import force_unicode as force_text
 except ImportError:
@@ -11,13 +12,26 @@ from django.utils.functional import allow_lazy
 from crispy_forms.compatibility import text_type
 
 register = template.Library()
+TEMPLATE_PACK = getattr(settings, 'CRISPY_TEMPLATE_PACK', 'bootstrap')
 
 
-def selectively_remove_spaces_between_tags(value):
-    html = re.sub(r'>\s+<', '><', force_text(value))
-    html = re.sub(r'</button><', '</button> <', force_text(html))
-    return re.sub(r'/><', r'/> <', force_text(html))
-selectively_remove_spaces_between_tags = allow_lazy(selectively_remove_spaces_between_tags, text_type)
+def selectively_remove_spaces_between_tags(value, template_pack, form_class):
+    if (
+        'bootstrap' in template_pack
+        and 'form-inline' in form_class
+    ):
+        # Bootstrap inline forms rely on spaces separating inputs, really
+        html = re.sub(r'>\s+<', '> <', force_text(value))
+        html = re.sub(r'</button><', '</button> <', force_text(html))
+        return re.sub(r'/><', r'/> <', force_text(html))
+    else:
+        html = re.sub(r'>\s+<', '><', force_text(value))
+        html = re.sub(r'</button><', '</button> <', force_text(html))
+        return re.sub(r'/><', r'/> <', force_text(html))
+    return value
+selectively_remove_spaces_between_tags = allow_lazy(
+    selectively_remove_spaces_between_tags, text_type
+)
 
 
 class SpecialSpacelessNode(template.Node):
@@ -25,8 +39,20 @@ class SpecialSpacelessNode(template.Node):
         self.nodelist = nodelist
 
     def render(self, context):
+        try:
+            template_pack = template.Variable('template_pack').resolve(context)
+        except:
+            template_pack = TEMPLATE_PACK
+
+        try:
+            form_attrs = template.Variable('form_attrs').resolve(context)
+        except:
+            form_attrs = {}
+
         return selectively_remove_spaces_between_tags(
-            self.nodelist.render(context).strip()
+            self.nodelist.render(context).strip(),
+            template_pack,
+            form_attrs.get('class', ''),
         )
 
 
@@ -39,4 +65,5 @@ def specialspaceless(parser, token):
     """
     nodelist = parser.parse(('endspecialspaceless',))
     parser.delete_first_token()
+
     return SpecialSpacelessNode(nodelist)